Warnings.

+++ tests/dsa_test.c
@@ -280,7 +280,7 @@ static int s_dsa_wycheproof_test(void)
    dsa_key key;
    int stat;
 
-   DO(base64_decode(b64key, XSTRLEN(b64key), derkey, &derlen));
+   DO(base64_decode(b64key, (unsigned long)XSTRLEN(b64key), derkey, &derlen));
    if (derlen != 838) {
       fprintf(stderr, "base64_decode failed, derlen=%lu (expected 838)\n", derlen);
       return CRYPT_FAIL_TESTVECTOR;
